replaced DynamicAvatar on UserPageHeader
