+++ react-app/src/components/SongFolders/EditSongForm/index.js
@@ -127,7 +127,7 @@ const EditSongForm = ({ setShowEditSongModal }) => {
       onClick: handleCancel,
       type: "cancel",
     },
-  ]
+  ];
 
   return (
     <>
