+++ react-app/src/components/UserPage/UserPageHeader.js
@@ -3,7 +3,9 @@ import { useDispatch, useSelector } from "react-redux";
 import { useParams } from "react-router-dom";
 import { editUserDetails } from "../../store/user";
 import DropdownButton from "../Buttons/DropdownButton";
-import DynamicAvatar from "./DynamicAvatar";
+import DynamicImage from "../DynamicImage";
+import { Modal } from "../Context/Modal";
+import ConfirmModal from "../ConfirmModal";
 
 const UserPageHeader = () => {
   const dispatch = useDispatch();
@@ -11,6 +13,8 @@ const UserPageHeader = () => {
   const sessionUser = useSelector(state => state.session.user);
   const user = useSelector(state => state.users[userId]);
   const [showDropdown, setShowDropdown] = useState(false);
+  const [showConfirmModal, setShowConfirmModal] = useState(false);
+  const avatarInputId = "avatar-url";
 
   useEffect(() => {
     if (!showDropdown) return;
@@ -24,6 +28,27 @@ const UserPageHeader = () => {
 
     return () => document.removeEventListener("click", closeDropdown);
   }, [showDropdown]);
+
+  const updateAvatar = async (e) => {
+    const file = e.target.files[0];
+
+    if (file) {
+      const formData = new FormData();
+      formData.append('avatar_url', file);
+      await dispatch(editUserDetails(+userId, formData));
+    }
+  };
+
+  const confirmDelete = e => {
+    setShowConfirmModal(true);
+  };
+
+  const deleteAvatarUrl = async (e) => {
+    const formData = new FormData();
+    formData.append('avatar_url', '');
+    await dispatch(editUserDetails(+userId, formData));
+    setShowConfirmModal(false);
+  };
 
   const deleteBannerUrl = async (e) => {
     const formData = new FormData();
@@ -78,9 +103,31 @@ const UserPageHeader = () => {
         className="banner-content flex-row"
       >
         <div className="user-page-banner-left flex-row">
-          <div style={{ height: "200px", width: "200px" }}>
-            <DynamicAvatar />
-          </div>
+          <DynamicImage
+            dimension={200}
+            standalone={true}
+            entity="user"
+            imageUrl={user?.avatar_url}
+            hiddenInput={
+              <input
+                type="file"
+                accept="image/*"
+                onChange={updateAvatar}
+                name={avatarInputId}
+                id={avatarInputId}
+                hidden
+              />
+            }
+            isAuthorized={sessionUser.id === +userId}
+            clickHidden={() => document.getElementById(avatarInputId).click()}
+            styleClasses={['button-action', 'b2']}
+            uploadLabel="Upload image"
+            replaceLabel="Replace image"
+            updateLabel="Update image"
+            deleteLabel="Delete image"
+            beforeLabel="camera-label"
+            confirmDelete={confirmDelete}
+          />
           <h2>{user?.display_name}</h2>
         </div>
         {sessionUser.id === +userId && (
@@ -114,6 +161,21 @@ const UserPageHeader = () => {
           </>
         )}
       </div>
+      {showConfirmModal && (
+        <Modal
+          onClose={() => setShowConfirmModal(false)}
+          position="top"
+        >
+          <ConfirmModal
+            setShowModal={setShowConfirmModal}
+            handleDelete={deleteAvatarUrl}
+            body={<p>
+              Please confirm that you want to delete this image.<br />
+              This action cannot be reversed.
+            </p>}
+          />
+        </Modal>
+      )}
     </header>
   );
 };
